+++ OUDS/Core/OUDS/Sources/OUDSTheme/OUDSTheme+SemanticTokens/OUDSTheme+ElevationMultipleSemanticTokens.swift
@@ -19,6 +19,8 @@ import OUDSTokensSemantic
 // WARNING: Not synchronized anymore with the Figjam / Figma by developers team
 // Create an issue for update https://github.com/Orange-OpenSource/ouds-ios/issues/new?template=token_update.yml
 
+// swiftlint:disable line_length
+
 /// Defines wrapper objects for eelvation color semantic  tokens.
 /// These values can be overriden inside ``OUDSTheme`` subclasses (in extensions or not, in the same module or not) thanks to the `@objc open` combination.
 extension OUDSTheme: ElevationMultipleSemanticTokens {
@@ -32,3 +34,5 @@ extension OUDSTheme: ElevationMultipleSemanticTokens {
     @objc open var elevationColorStickyEmphasized: MultipleColorSemanticTokens { MultipleColorSemanticTokens(light: elevationColorStickyEmphasizedLight, dark: elevationColorStickyEmphasizedDark) }
     @objc open var elevationColorStickyNavigationScrolled: MultipleColorSemanticTokens { MultipleColorSemanticTokens(light: elevationColorStickyNavigationScrolledLight, dark: elevationColorStickyNavigationScrolledDark) }
 }
+
+// swiftlint:enable line_length