+++ OUDS/Core/Tokens/RawTokens/Tests/ColorRawTokensTests.swift
@@ -260,83 +260,195 @@ final class ColorRawTokensTests: XCTestCase {
         XCTAssertColorLighterThan(ColorRawTokens.colorFunctionalDodgerBlue800, ColorRawTokens.colorFunctionalDodgerBlue900)
     }
 
-    // MARK: - Primitive token - Colors - Transparent black
+    // MARK: - Primitive token - Colors - Opacity black
 
-    func testColorTransparentBlack0LighterThanTransparentBlack100() throws {
-        XCTAssertColorDarkerThan(ColorRawTokens.colorTransparentBlack0, ColorRawTokens.colorTransparentBlack100)
+    func testColorOpacityBlack0LighterThanTransparentBlack40() throws {
+        XCTAssertColorDarkerThan(ColorRawTokens.colorOpacityBlack0, ColorRawTokens.colorOpacityBlack40)
     }
 
-    func testColorTransparentBlack100LighterThanTransparentBlack200() throws {
-        XCTAssertColorDarkerThan(ColorRawTokens.colorTransparentBlack100, ColorRawTokens.colorTransparentBlack200)
+    func testColorOpacityBlack40LighterThanTransparentBlack80() throws {
+        XCTAssertColorDarkerThan(ColorRawTokens.colorOpacityBlack40, ColorRawTokens.colorOpacityBlack80)
     }
 
-    func testColorTransparentBlack200LighterThanTransparentBlack300() throws {
-        XCTAssertColorDarkerThan(ColorRawTokens.colorTransparentBlack200, ColorRawTokens.colorTransparentBlack300)
+    func testColorOpacityBlack80LighterThanTransparentBlack120() throws {
+        XCTAssertColorDarkerThan(ColorRawTokens.colorOpacityBlack80, ColorRawTokens.colorOpacityBlack120)
     }
 
-    func testColorTransparentBlack300LighterThanTransparentBlack400() throws {
-        XCTAssertColorDarkerThan(ColorRawTokens.colorTransparentBlack300, ColorRawTokens.colorTransparentBlack400)
+    func testColorOpacityBlack120LighterThanTransparentBlack160() throws {
+        XCTAssertColorDarkerThan(ColorRawTokens.colorOpacityBlack120, ColorRawTokens.colorOpacityBlack160)
     }
 
-    func testColorTransparentBlack400LighterThanTransparentBlack500() throws {
-        XCTAssertColorDarkerThan(ColorRawTokens.colorTransparentBlack400, ColorRawTokens.colorTransparentBlack500)
+    func testColorOpacityBlack160LighterThanTransparentBlack200() throws {
+        XCTAssertColorDarkerThan(ColorRawTokens.colorOpacityBlack160, ColorRawTokens.colorOpacityBlack200)
     }
 
-    func testColorTransparentBlack500LighterThanTransparentBlack600() throws {
-        XCTAssertColorDarkerThan(ColorRawTokens.colorTransparentBlack500, ColorRawTokens.colorTransparentBlack600)
+    func testColorOpacityBlack200LighterThanTransparentBlack240() throws {
+        XCTAssertColorDarkerThan(ColorRawTokens.colorOpacityBlack200, ColorRawTokens.colorOpacityBlack240)
     }
 
-    func testColorTransparentBlack600LighterThanTransparentBlack700() throws {
-        XCTAssertColorDarkerThan(ColorRawTokens.colorTransparentBlack600, ColorRawTokens.colorTransparentBlack700)
+    func testColorOpacityBlack240LighterThanTransparentBlack280() throws {
+        XCTAssertColorDarkerThan(ColorRawTokens.colorOpacityBlack240, ColorRawTokens.colorOpacityBlack280)
     }
 
-    func testColorTransparentBlack700LighterThanTransparentBlack800() throws {
-        XCTAssertColorDarkerThan(ColorRawTokens.colorTransparentBlack700, ColorRawTokens.colorTransparentBlack800)
+    func testColorOpacityBlack280LighterThanTransparentBlack320() throws {
+        XCTAssertColorDarkerThan(ColorRawTokens.colorOpacityBlack280, ColorRawTokens.colorOpacityBlack320)
     }
 
-    func testColorTransparentBlack800LighterThanTransparentBlack900() throws {
-        XCTAssertColorDarkerThan(ColorRawTokens.colorTransparentBlack800, ColorRawTokens.colorTransparentBlack900)
+    func testColorOpacityBlack320LighterThanTransparentBlack360() throws {
+        XCTAssertColorDarkerThan(ColorRawTokens.colorOpacityBlack320, ColorRawTokens.colorOpacityBlack360)
     }
 
-    // MARK: - Primitive token - Colors - Transparent white
-
-    func testColorRawTokenColorTransparentWhite0LighterThanColorTransparentWhite100() throws {
-        XCTAssertColorDarkerThan(ColorRawTokens.colorTransparentWhite0, ColorRawTokens.colorTransparentWhite100)
+    func testColorOpacityBlack360LighterThanTransparentBlack400() throws {
+        XCTAssertColorDarkerThan(ColorRawTokens.colorOpacityBlack360, ColorRawTokens.colorOpacityBlack400)
     }
 
-    func testColorRawTokenColorTransparentWhite100LighterThanColorTransparentWhite200() throws {
-        XCTAssertColorDarkerThan(ColorRawTokens.colorTransparentWhite100, ColorRawTokens.colorTransparentWhite200)
+    func testColorOpacityBlack400LighterThanTransparentBlack440() throws {
+        XCTAssertColorDarkerThan(ColorRawTokens.colorOpacityBlack400, ColorRawTokens.colorOpacityBlack440)
     }
 
-    func testColorRawTokenColorTransparentWhite200LighterThanColorTransparentWhite300() throws {
-        XCTAssertColorDarkerThan(ColorRawTokens.colorTransparentWhite200, ColorRawTokens.colorTransparentWhite300)
+    func testColorOpacityBlack440LighterThanTransparentBlack480() throws {
+        XCTAssertColorDarkerThan(ColorRawTokens.colorOpacityBlack440, ColorRawTokens.colorOpacityBlack480)
     }
 
-    func testColorRawTokenColorTransparentWhite300LighterThanColorTransparentWhite400() throws {
-        XCTAssertColorDarkerThan(ColorRawTokens.colorTransparentWhite300, ColorRawTokens.colorTransparentWhite400)
+    func testColorOpacityBlack480LighterThanTransparentBlack520() throws {
+        XCTAssertColorDarkerThan(ColorRawTokens.colorOpacityBlack480, ColorRawTokens.colorOpacityBlack520)
     }
 
-    func testColorRawTokenColorTransparentWhite400LighterThanColorTransparentWhite500() throws {
-        XCTAssertColorDarkerThan(ColorRawTokens.colorTransparentWhite400, ColorRawTokens.colorTransparentWhite500)
+    func testColorOpacityBlack520LighterThanTransparentBlack560() throws {
+        XCTAssertColorDarkerThan(ColorRawTokens.colorOpacityBlack520, ColorRawTokens.colorOpacityBlack560)
     }
 
-    func testColorRawTokenColorTransparentWhite500LighterThanColorTransparentWhite600() throws {
-        XCTAssertColorDarkerThan(ColorRawTokens.colorTransparentWhite500, ColorRawTokens.colorTransparentWhite600)
+    func testColorOpacityBlack560LighterThanTransparentBlack600() throws {
+        XCTAssertColorDarkerThan(ColorRawTokens.colorOpacityBlack560, ColorRawTokens.colorOpacityBlack600)
     }
 
-    func testColorRawTokenColorTransparentWhite600LighterThanColorTransparentWhite700() throws {
-        XCTAssertColorDarkerThan(ColorRawTokens.colorTransparentWhite600, ColorRawTokens.colorTransparentWhite700)
+    func testColorOpacityBlack600LighterThanTransparentBlack640() throws {
+        XCTAssertColorDarkerThan(ColorRawTokens.colorOpacityBlack600, ColorRawTokens.colorOpacityBlack640)
     }
 
-    func testColorRawTokenColorTransparentWhite700LighterThanColorTransparentWhite800() throws {
-        XCTAssertColorDarkerThan(ColorRawTokens.colorTransparentWhite700, ColorRawTokens.colorTransparentWhite800)
+    func testColorOpacityBlack640LighterThanTransparentBlack680() throws {
+        XCTAssertColorDarkerThan(ColorRawTokens.colorOpacityBlack640, ColorRawTokens.colorOpacityBlack680)
     }
 
-    func testColorRawTokenColorTransparentWhite800LighterThanColorTransparentWhite900() throws {
-        XCTAssertColorDarkerThan(ColorRawTokens.colorTransparentWhite800, ColorRawTokens.colorTransparentWhite900)
+    func testColorOpacityBlack680LighterThanTransparentBlack720() throws {
+        XCTAssertColorDarkerThan(ColorRawTokens.colorOpacityBlack680, ColorRawTokens.colorOpacityBlack720)
     }
 
-    // MARK: Primitive token - Colors -  Emerald
+    func testColorOpacityBlack720LighterThanTransparentBlack760() throws {
+        XCTAssertColorDarkerThan(ColorRawTokens.colorOpacityBlack720, ColorRawTokens.colorOpacityBlack760)
+    }
+
+    func testColorOpacityBlack760LighterThanTransparentBlack800() throws {
+        XCTAssertColorDarkerThan(ColorRawTokens.colorOpacityBlack760, ColorRawTokens.colorOpacityBlack800)
+    }
+
+    func testColorOpacityBlack800LighterThanTransparentBlack840() throws {
+        XCTAssertColorDarkerThan(ColorRawTokens.colorOpacityBlack800, ColorRawTokens.colorOpacityBlack840)
+    }
+
+    func testColorOpacityBlack880LighterThanTransparentBlack920() throws {
+        XCTAssertColorDarkerThan(ColorRawTokens.colorOpacityBlack880, ColorRawTokens.colorOpacityBlack920)
+    }
+
+    func testColorOpacityBlack920LighterThanTransparentBlack960() throws {
+        XCTAssertColorDarkerThan(ColorRawTokens.colorOpacityBlack920, ColorRawTokens.colorOpacityBlack960)
+    }
+
+    // MARK: - Primitive token - Colors - Opacity White
+
+    func testColorOpacityWhite0LighterThanTransparentWhite40() throws {
+        XCTAssertColorDarkerThan(ColorRawTokens.colorOpacityWhite0, ColorRawTokens.colorOpacityWhite40)
+    }
+
+    func testColorOpacityWhite40LighterThanTransparentWhite80() throws {
+        XCTAssertColorDarkerThan(ColorRawTokens.colorOpacityWhite40, ColorRawTokens.colorOpacityWhite80)
+    }
+
+    func testColorOpacityWhite80LighterThanTransparentWhite120() throws {
+        XCTAssertColorDarkerThan(ColorRawTokens.colorOpacityWhite80, ColorRawTokens.colorOpacityWhite120)
+    }
+
+    func testColorOpacityWhite120LighterThanTransparentWhite160() throws {
+        XCTAssertColorDarkerThan(ColorRawTokens.colorOpacityWhite120, ColorRawTokens.colorOpacityWhite160)
+    }
+
+    func testColorOpacityWhite160LighterThanTransparentWhite200() throws {
+        XCTAssertColorDarkerThan(ColorRawTokens.colorOpacityWhite160, ColorRawTokens.colorOpacityWhite200)
+    }
+
+    func testColorOpacityWhite200LighterThanTransparentWhite240() throws {
+        XCTAssertColorDarkerThan(ColorRawTokens.colorOpacityWhite200, ColorRawTokens.colorOpacityWhite240)
+    }
+
+    func testColorOpacityWhite240LighterThanTransparentWhite280() throws {
+        XCTAssertColorDarkerThan(ColorRawTokens.colorOpacityWhite240, ColorRawTokens.colorOpacityWhite280)
+    }
+
+    func testColorOpacityWhite280LighterThanTransparentWhite320() throws {
+        XCTAssertColorDarkerThan(ColorRawTokens.colorOpacityWhite280, ColorRawTokens.colorOpacityWhite320)
+    }
+
+    func testColorOpacityWhite320LighterThanTransparentWhite360() throws {
+        XCTAssertColorDarkerThan(ColorRawTokens.colorOpacityWhite320, ColorRawTokens.colorOpacityWhite360)
+    }
+
+    func testColorOpacityWhite360LighterThanTransparentWhite400() throws {
+        XCTAssertColorDarkerThan(ColorRawTokens.colorOpacityWhite360, ColorRawTokens.colorOpacityWhite400)
+    }
+
+    func testColorOpacityWhite400LighterThanTransparentWhite440() throws {
+        XCTAssertColorDarkerThan(ColorRawTokens.colorOpacityWhite400, ColorRawTokens.colorOpacityWhite440)
+    }
+
+    func testColorOpacityWhite440LighterThanTransparentWhite480() throws {
+        XCTAssertColorDarkerThan(ColorRawTokens.colorOpacityWhite440, ColorRawTokens.colorOpacityWhite480)
+    }
+
+    func testColorOpacityWhite480LighterThanTransparentWhite520() throws {
+        XCTAssertColorDarkerThan(ColorRawTokens.colorOpacityWhite480, ColorRawTokens.colorOpacityWhite520)
+    }
+
+    func testColorOpacityWhite520LighterThanTransparentWhite560() throws {
+        XCTAssertColorDarkerThan(ColorRawTokens.colorOpacityWhite520, ColorRawTokens.colorOpacityWhite560)
+    }
+
+    func testColorOpacityWhite560LighterThanTransparentWhite600() throws {
+        XCTAssertColorDarkerThan(ColorRawTokens.colorOpacityWhite560, ColorRawTokens.colorOpacityWhite600)
+    }
+
+    func testColorOpacityWhite600LighterThanTransparentWhite640() throws {
+        XCTAssertColorDarkerThan(ColorRawTokens.colorOpacityWhite600, ColorRawTokens.colorOpacityWhite640)
+    }
+
+    func testColorOpacityWhite640LighterThanTransparentWhite680() throws {
+        XCTAssertColorDarkerThan(ColorRawTokens.colorOpacityWhite640, ColorRawTokens.colorOpacityWhite680)
+    }
+
+    func testColorOpacityWhite680LighterThanTransparentWhite720() throws {
+        XCTAssertColorDarkerThan(ColorRawTokens.colorOpacityWhite680, ColorRawTokens.colorOpacityWhite720)
+    }
+
+    func testColorOpacityWhite720LighterThanTransparentWhite760() throws {
+        XCTAssertColorDarkerThan(ColorRawTokens.colorOpacityWhite720, ColorRawTokens.colorOpacityWhite760)
+    }
+
+    func testColorOpacityWhite760LighterThanTransparentWhite800() throws {
+        XCTAssertColorDarkerThan(ColorRawTokens.colorOpacityWhite760, ColorRawTokens.colorOpacityWhite800)
+    }
+
+    func testColorOpacityWhite800LighterThanTransparentWhite840() throws {
+        XCTAssertColorDarkerThan(ColorRawTokens.colorOpacityWhite800, ColorRawTokens.colorOpacityWhite840)
+    }
+
+    func testColorOpacityWhite880LighterThanTransparentWhite920() throws {
+        XCTAssertColorDarkerThan(ColorRawTokens.colorOpacityWhite880, ColorRawTokens.colorOpacityWhite920)
+    }
+
+    func testColorOpacityWhite920LighterThanTransparentWhite960() throws {
+        XCTAssertColorDarkerThan(ColorRawTokens.colorOpacityWhite920, ColorRawTokens.colorOpacityWhite960)
+    }
+
+    // MARK: Primitive token - Colors - Emerald
 
     func testOrangeBrandColorRawTokenColorDecorativeEmerald100LighterThanEmerald200() throws {
         XCTAssertColorLighterThan(ColorRawTokens.colorDecorativeEmerald100, ColorRawTokens.colorDecorativeEmerald200)
@@ -404,7 +516,7 @@ final class ColorRawTokensTests: XCTestCase {
         XCTAssertColorLighterThan(ColorRawTokens.colorDecorativeSky800, ColorRawTokens.colorDecorativeSky900)
     }
 
-    // MARK: Primitive token - Colors -  Amber
+    // MARK: Primitive token - Colors - Amber
 
     func testOrangeBrandColorRawTokenColorDecorativeAmber100LighterThanAmber200() throws {
         XCTAssertColorLighterThan(ColorRawTokens.colorDecorativeAmber100, ColorRawTokens.colorDecorativeAmber200)
@@ -438,7 +550,7 @@ final class ColorRawTokensTests: XCTestCase {
         XCTAssertColorLighterThan(ColorRawTokens.colorDecorativeAmber800, ColorRawTokens.colorDecorativeAmber900)
     }
 
-    // MARK: Primitive token - Colors -  Amethyst
+    // MARK: Primitive token - Colors - Amethyst
 
     func testOrangeBrandColorRawTokenColorDecorativeAmethyst100LighterThanAmethyst200() throws {
         XCTAssertColorLighterThan(ColorRawTokens.colorDecorativeAmethyst100, ColorRawTokens.colorDecorativeAmethyst200)